version 92

+++ scripts/install-ssmtp.sh
@@ -2,22 +2,6 @@
 set -u
 set -e
 
-echo -e "
-Are you sure you want to install $(tput bold)ssmtp$(tput sgr0)? This will
-do the following tasks:
-
-  - install $(tput setab 5)ssmtp$(tput sgr0) packages
-  - replace your ssmtp.config (if you want to)
-  - set a samba password for current user (if you want to)
-"
-read -r -p "Do you want to proceed? (y/n)? " INSTALL_CONTROL
-[[ "${INSTALL_CONTROL}" == "y" ]] || exit 0
-
-IS_REPOSITORY_URL="https://raw.githubusercontent.com/lukasdanckwerth/notes/main"
-IS_DEFAULT_CONFIG_URL="${IS_REPOSITORY_URL}/assets/ssmtp/ssmtp.conf"
-IS_SSMTP_CONFIG="/etc/ssmtp/ssmtp.conf"
-IS_SSMTP_CONFIG_TEMP="/tmp/install-ssmtp.sh-ssmtp.conf"
-
 log() {
   echo -e "[install-ssmtp]  ${*}"
 }
@@ -30,18 +14,43 @@ die() {
   log "$(tput setaf 9)ERROR: ${*} $(tput sgr0)" && exit 1
 }
 
+[[ "$(whoami)" == "root" ]] || die "script must be run as root"
+
+echo -e "
+Are you sure you want to install $(tput bold)ssmtp$(tput sgr0)? This will
+do the following tasks:
+
+  - install $(tput setab 5)ssmtp mailutils$(tput sgr0) packages
+  - replace your ssmtp.config (if you want to)
+"
+read -r -p "Do you want to proceed? (y/n)? " INSTALL_CONTROL
+[[ "${INSTALL_CONTROL}" == "y" ]] || exit 0
+
+IS_REPOSITORY_URL="https://raw.githubusercontent.com/lukasdanckwerth/notes/main"
+IS_DEFAULT_CONFIG_URL="${IS_REPOSITORY_URL}/assets/ssmtp/ssmtp.conf"
+
+IS_SSMTP_CONFIG="/etc/ssmtp/ssmtp.conf"
+IS_SSMTP_CONFIG_TEMP="/tmp/install-ssmtp.sh-ssmtp.conf"
+
+IS_SSMTP_REVALIASES="/etc/ssmtp/revaliases"
+
+env
+exit 0
+
 log "start"
 log "install packages"
 sudo apt-get install --assume-yes ssmtp mailutils
 
 [[ -d "/etc/ssmtp" ]] || die "directory /etc/ssmtp not existing."
 
+echo
 read -r -p "Do you want to run the wizard? (y/n) " replaceConfig
 
 # replace ssmtp.conf
 if [[ "${replaceConfig}" == "y" ]]; then
+
   log "download ssmtp.conf to ${IS_SSMTP_CONFIG_TEMP}"
-  curl "${IS_DEFAULT_CONFIG_URL}" -o "${IS_SSMTP_CONFIG_TEMP}"
+  curl --silent "${IS_DEFAULT_CONFIG_URL}" -o "${IS_SSMTP_CONFIG_TEMP}"
 
   [[ -f "${IS_SSMTP_CONFIG_TEMP}" ]] || die "couldn't load ssmtp.conf ${IS_SSMTP_CONFIG}"
 
@@ -52,34 +61,32 @@ if [[ "${replaceConfig}" == "y" ]]; then
   sudo mv "${IS_SSMTP_CONFIG_TEMP}" "${IS_SSMTP_CONFIG}"
 
   [[ -f "${IS_SSMTP_CONFIG}" ]] || die "couldn't move ${IS_SSMTP_CONFIG}"
+  sudo chown root:mail "${IS_SSMTP_CONFIG}"
+  sudo chmod 640 "${IS_SSMTP_CONFIG}"
 
-  log "${IS_SSMTP_CONFIG} file"
-  sudo ls -la "${IS_SSMTP_CONFIG}"
+  log "$(sudo ls -la "${IS_SSMTP_CONFIG}")"
 
-  log "${IS_SSMTP_CONFIG} content"
-  log ""
-  sudo cat "${IS_SSMTP_CONFIG}"
-  log ""
+  log "Please provide a mailhub"
+  log "  - mail.gmx.net:465"
 
-  echo -e "
-  Please provide a mailhub.
+  read -r -p "Enter mailhub (enter for GMX): " SM_MAILHUB_TEMP
+  [[ "${SM_MAILHUB_TEMP}" == "" ]] && SM_MAILHUB="mail.gmx.net:465"
+  [[ "${SM_MAILHUB_TEMP}" == "" ]] || SM_MAILHUB="${SM_MAILHUB_TEMP}"
 
-  - GMX mail.gmx.net:465
-  "
-  read -r -p "Enter mailhub: " SM_MAILHUB
   read -r -p "Enter mail adress: " SM_MAIL
 
   SM_PASSWORD_1="1"
   SM_PASSWORD_2="2"
 
-  while [[ ! "${SM_PASSWORD_1}" == "${SM_PASSWORD_1}" ]]; do
-    read -r -p "Enter password: " SM_PASSWORD_1
-    read -r -p "Repeat password (again): " SM_PASSWORD_2
+  while [[ ! "${SM_PASSWORD_1}" == "${SM_PASSWORD_2}" ]]; do
+    read -s -r -p "Enter password: " SM_PASSWORD_1
+    echo
+    read -s -r -p "Repeat password (again): " SM_PASSWORD_2
   done
 
   echo "mailhub:   ${SM_MAILHUB}"
   echo "mail:      ${SM_MAIL}"
-  echo "password:  ****"
+  echo "password:  $(echo "${SM_PASSWORD_1}" | wc -c | xargs) characters"
   read -r -p "Is this correct (y/n)? " IS_DATA_CORRECT
 
   [[ "${IS_DATA_CORRECT}" == "y" ]] || exit 0
@@ -87,6 +94,18 @@ if [[ "${replaceConfig}" == "y" ]]; then
   sed -i "s/__MAILHUB__/${SM_MAILHUB}/g" "${IS_SSMTP_CONFIG}"
   sed -i "s/__MAIL__/${SM_MAIL}/g" "${IS_SSMTP_CONFIG}"
   sed -i "s/__PASSWORD__/${SM_PASSWORD_1}/g" "${IS_SSMTP_CONFIG}"
+
+  read -r -p "Update ${IS_SSMTP_REVALIASES} (y/n)? " IS_UPDATE_REVALIASES
+  [[ "${IS_UPDATE_REVALIASES}" == "y" ]] || exit 0
+  sudo rm -rf "${IS_SSMTP_REVALIASES}"
+  sudo touch "${IS_SSMTP_REVALIASES}"
+
+  read -r -p "Use newly configured mail account for $(bold "root") (y/n)? " IS_USE_MAIL_ROOT
+  [[ "${IS_USE_MAIL_ROOT}" == "y" ]] && echo "root:${SM_MAIL}:${SM_MAILHUB}" >> ${IS_SSMTP_REVALIASES}
+
+  IS_SUDO_USER=$(SUDO_USER)
+  read -r -p "Use newly configured mail account for $(bold "${IS_SUDO_USER}") (y/n)? " IS_USE_MAIL_USER
+  [[ "${IS_USE_MAIL_USER}" == "y" ]] && echo "${IS_SUDO_USER}:${SM_MAIL}:${SM_MAILHUB}" >> ${IS_SSMTP_REVALIASES}
 
 else
   read -r -p "Do you want to edit the ssmtp.conf right now (y/n)? " SET_PASSWORD
